pyproject.toml
[project]
name = "scaling-llms"
version = "0.1.0"
description = ""
authors = [
    {name = "Vasilis Stylianou",email = "stylianouvasilis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "ipykernel (>=7.1.0,<8.0.0)",
    "numpy (<2)",
    "pandas (<3)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "tensorboard (>=2.20.0)",
    "notebook (>=7.5.3,<8.0.0)",
    "datasets (>=4.5.0,<5.0.0)",
    "tiktoken (>=0.7.0,<1.0.0)",
    "torch (==2.2.2)",
    "torchvision (==0.17.2)",
    "torchaudio (==2.2.2)",
    "ipywidgets (>=8.1.8,<9.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "runpod (>=1.8.1,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
